migrated to react ts (#34)
