+++ src/App.tsx
@@ -1,0 +1,11 @@
+import './App.css'
+
+function App() {
+  return (
+    <>
+    <h2>Hello world</h2>
+    </>
+  );
+}
+
+export default App;
